geotiff tests: consolidate remote-hardening cluster (#2436)
Fold the five top-level remote-hardening files into
integration/test_http_sources.py, one Section per source issue:

- test_ssrf_hardening_1664.py
- test_dns_rebinding_pin_issue_1846.py
- test_uppercase_scheme_ssrf_2323.py
- test_max_cloud_bytes_dispatcher_silent_drop_2026_05_15.py
- test_open_geotiff_max_cloud_bytes_annot_2106.py

The #1846 file imported the redirect-mock pool from the #1664 file; that
cross-file import becomes an in-file reference now that both live in the
same module. Colliding helpers are namespaced per source issue and the
per-file cupy probe collapses onto the shared requires_gpu marker. No
assertion changed; 74 test functions (107 collected) match the
pre-consolidation total.

The release-gate checklist and geotiff docs cite the new path so the
checklist-parity gate stays green.

Sub-PR B of issue #2436 (cluster 12, long-tail epic #2424). Tests-only.
